+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 ### Fixed
+- Queries which have been removed Flowmachine's cache, or cancelled can now be rerun. [#1898](https://github.com/Flowminder/FlowKit/issues/1898)
 
 ### Removed